Moved server test to test folder
